Add version identification to firmware image

+++ do_bootfs.sh
@@ -1,5 +1,12 @@
 #!/bin/sh 
 
+VERSION=
+
+if [ x${VERSION} = x ] ; then
+   echo Set a proper version string
+   exit
+   fi
+
 dd if=/dev/zero of=initrd.img bs=1M count=3 
 mke2fs -F -b 1024 -i 1024 initrd.img 
 tune2fs -c 0 -i 0 initrd.img 
@@ -8,6 +15,7 @@ mkdir image
 mount -o loop -t auto initrd.img image 
 tar cC bootfs . | tar xC image 
 tar xC image -f bootfs-skel.tar
+echo ${VERSION} > image/etc/sysconfig/config/version
 echo 'ttyS0::respawn:/sbin/getty -L ttyS0 115200 vt100' >> image/etc/inittab
 umount image 
 rmdir image
